ref(perf): Move transaction vitals under transaction summary As we add more subpages (eg. tags page) it will get confusing unless we group views closer to how the routes are structured.
