+++ static/app/views/performance/transactionSummary/transactionVitals/vitalsPanel.tsx
