+++ tests/js/spec/views/performance/transactionVitals/utils.spec.jsx
@@ -2,7 +2,7 @@ import {
   findNearestBucketIndex,
   getRefRect,
   mapPoint,
-} from 'app/views/performance/transactionVitals/utils';
+} from 'app/views/performance/transactionSummary/transactionVitals/utils';
 
 describe('Utils', function () {
   describe('findNearestBucketIndex()', function () {
